requirements.txt
Django==3.2
djangorestframework
psycopg2-binary==2.8.6
Pillow
django-filter
python-decouple
drf-yasg
selenium
djoser
django-rest-framework-social-oauth2

